requirements.txt
tensorflow-cpu
albumentations
einops
fastapi[standard]
numpy
pandas
seaborn
tqdm
matplotlib
datasets
opencv-python
python-box
awscli
joblib
dvc
tensorflow-serving-api
py-grpc-prometheus